requirements.txt
requests
xmltodict
pytest
selenium
allure-pytest
webdriver_manager